app/requirements.txt
streamlit
fastai==1.0.57
torch==1.2.0
python-docx


